+++ src/main/java/com/learningspring/bookStore/service/BookStoreServiceImpl.java
@@ -43,8 +43,13 @@ public class BookStoreServiceImpl implements BookStoreService {
     }
 
     @Override
-    public void deleteBook(Long id) {
-        bookRepository.deleteById(id);
+    public void deleteBook(Long id) throws ResourceNotFoundException {
+
+        Optional<Book> book = bookRepository.findById(id);
+        if(book.isPresent()) {
+            bookRepository.deleteById(id);
+        }
+        else throw new ResourceNotFoundException("Book is not available in store:" + id);
     }
 
 }
